Huge Optimisation!
previously every single step, the ghost script would recheck the validity of the ghost block, which cycled through every single tile of the tetromino, in every space between the active tetromino, and the floor. This would happen even if the active tetromino didnt move. By adding a simple check, it now only checks the ghost validity if the active piece has sinced moved on the X axis, or rotated, this saves a lot of needless loops.