Small bug fix
Fixed a bug where the I tetromino ghost didn't always touch the floor when on it's side in a specific rotation position. Added another row check so that the loop didn't premeturely end (which was what was happening). Fix doesn't affect other pieces as IsValidPosition would just quickly return null.