Code Optimisation
Previously the IsValidPosition method would wait until it detected both outOfBounds AND colliding before quitting. There's no real uses for needing both of this information, as whenether the piece is outOfBounds, its invalid. Changed the code to break after either an outOfBounds OR colliding is detected, which potentially saves a lot of tile checks.