Initial AI Commit
Initial commit containing working rule-based AI player. The AI is too poor so loses the game quite quickly, it is also so fast that I implemented a button which takes control each new time the button is selected, until the current piece is locked in. I tried to implement a delay to the AI movement so that the button isn't needed but was unsuccessful