Tetronimo Spawn Bug Fix
As the tetromino spawn position is above the play area so that it falls into view, they were visible outside of the play area. I temporarily placed a box to hide it - I've now added a check before a tile is drawn, fixing this issue. I also moved the highscore text as it clipped with the wall if it was too high.